Improved REAADME and formatting
