+++ samples/SampleLibrary.HandRolledTests/SampleTest.cs
@@ -1,8 +1,8 @@
+using System;
+using System.Collections.Generic;
 using Microsoft.Extensions.Logging;
 using SampleLibrary;
 using Xunit;
-using System.Collections.Generic;
-using System;
 
 #nullable disable
