+++ samples/SampleWebApplication.IntegrationTests/LoggingTest.cs
@@ -1,8 +1,8 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.Logging;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace SampleWebApplication.IntegrationTests
