+++ samples/SampleWebApplication2_1.IntegrationTests/LoggingTestWithInjectedFactory.cs
@@ -1,6 +1,6 @@
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace SampleWebApplication2_1.IntegrationTests
