+++ samples/SampleWebApplicationSerilog.IntegrationTests/LoggingTest.cs
@@ -1,9 +1,9 @@
+using System.Threading.Tasks;
 using MELT;
 using MELT.Xunit;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.Logging;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace SampleWebApplicationSerilog.Tests
