+++ samples/SampleWebApplicationSerilog/Program.cs
@@ -1,8 +1,8 @@
+using System;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 using Serilog;
 using Serilog.Events;
-using System;
 
 namespace SampleWebApplicationSerilog
 {
