--- src/MELT.Serilog.Xunit/MELT.Serilog.Xunit.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <Description>xUnit helper for the MELT testing library to verify that the correct logs are generated by apps or libraries using the .NET Standard Microsoft.Extensions.Logging library.</Description>
-    <PackageTags>$(PackageTags);xUnit</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="xunit.abstractions" Version="2.0.2" />
-    <PackageReference Include="xunit.assert" Version="2.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MELT.Serilog\MELT.Serilog.csproj" />
-  </ItemGroup>
-
-</Project>
