+++ src/MELT.Xunit/LoggingAssert.cs
@@ -1,4 +1,4 @@
-﻿// Copyright(c) .NET Foundation. All rights reserved.
+// Copyright(c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
