+++ src/MELT/Helpers/ITestSink.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace MELT
 {
-    #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
     [Obsolete]
     public interface ITestSink : IInternalTestSink, ITestLoggerSink
     {
