+++ src/MELT/Helpers/LogEntry.cs
@@ -1,6 +1,6 @@
-using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
+using Microsoft.Extensions.Logging;
 
 namespace MELT
 {
