+++ src/MELT/Helpers/MELTBuilder.cs
@@ -1,6 +1,6 @@
+using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using System;
 
 namespace MELT
 {
