+++ src/MELT/Helpers/TestSinkOptions.cs
@@ -1,5 +1,5 @@
-using Microsoft.Extensions.Logging;
 using System;
+using Microsoft.Extensions.Logging;
 
 namespace MELT
 {
