+++ test/MELT.Xunit.Tests/LogValuesAssertTest.cs
@@ -1,9 +1,9 @@
 // Copyright(c) .NET Foundation.All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using MELT.Xunit;
 using System.Collections.Generic;
 using System.Linq;
+using MELT.Xunit;
 using Xunit;
 using Xunit.Sdk;
